backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
openai-whisper==20231117
huggingface_hub>=0.22.0
transformers>=4.40.0
tokenizers>=0.19.0
soundfile==0.12.1
numpy>=1.24.0
scipy>=1.12.0
httpx==0.27.0
aiofiles==23.2.1
google-genai